pretty_fits/requirements.txt
git+git://github.com/LCOGT/fits2image.git#egg=fits2image
astropy
astroscrappy
Cython
numpy
Pillow
scipy
